+++ Changelog.md
@@ -1,5 +1,19 @@
 # Changelog
 
+**v1.7.0 (not yet released)**
+
+* Upgraded to Monogame 3.7
+    * fixes MSAA in DirectX based monogame
+* WpfGraphicsDeviceService now properly calls all IGraphicsDeviceService events
+* added support for DPI scaling on WpfGraphicsDeviceManager
+* by default, each WpfGame now has its own GraphicsDevice instance
+    * to enable old behaviour of one global GraphicsDevice instance: call WpfGame.UseASingleSharedGraphicsDevice = true **before the first WpfGame instance is created**)
+
+TODOs:
+
+* check amount of calls to events of WpfGraphicsDeviceService (currently they are called like 3x on initial window create)
+* check if wpf designer is working or crashing
+
 **v1.6.0**
 
 Switch to Monogame 3.6 and content builder nuget (removes the dependency to install monogame locally).